backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Legal Document RAG System with Citation Tracking"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.1.0",
    "langchain-ollama>=0.1.0",
    "langchain-community>=0.0.13",
    "neo4j>=5.14.1",
    "supabase>=2.3.0",
    "python-multipart>=0.0.6",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "tiktoken>=0.5.2",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "langchain-classic>=1.0.0",
]
